+++ app/logic/search.py
@@ -3,7 +3,7 @@ from app.models.laborStatusForm import LaborStatusForm
 from app.models.department import Department
 from app.models.formHistory import FormHistory
 
-def limitSearch(students):
+def limitSearch(students, currentUser):
     newstudents = []
     departments = list(FormHistory.select()
                     .join_from(FormHistory, LaborStatusForm)
@@ -11,8 +11,7 @@ def limitSearch(students):
                     .where((FormHistory.formID.supervisor == currentUser.supervisor.ID) | (FormHistory.createdBy == currentUser))
                     .distinct()
                     )
-    # for d in departments:
-    #     print(d.formID.department.DEPT_NAME)
+
     for student in students:
         student_in_department = (FormHistory.select()
                     .join_from(FormHistory, LaborStatusForm)
@@ -27,3 +26,15 @@ def limitSearch(students):
             newstudents.append(student)
 
             return newstudents
+
+def usernameFromEmail(email):
+    # split always returns a list, even if there is nothing to split, so [0] is safe
+    return email.split('@',1)[0]
+
+# Convert a Student or STUDATA record into the dictionary that our js expects
+def studentDbToDict(item):
+    return {'username': usernameFromEmail(item.STU_EMAIL.strip()),
+            'firstName': item.FIRST_NAME.strip(),
+            'lastName': item.LAST_NAME.strip(),
+            'bnumber': item.ID.strip(),
+            'type': 'Student'}
